+++ src/models/test/models_test.ts
@@ -280,6 +280,26 @@ describe('GenerativeModel startChat', () => {
 
     expect(chat).toBeInstanceOf(ChatSession);
   });
+  it('returns ChatSession when pass an arg with generationConfig including responseMimeType', async () => {
+    const fetchResult = Promise.resolve(
+      new Response(JSON.stringify(TEST_MODEL_RESPONSE), fetchResponseObj)
+    );
+    const fetchSpy = spyOn(global, 'fetch').and.returnValue(fetchResult);
+    const model = new GenerativeModel({
+      model: 'gemini-pro',
+      project: PROJECT,
+      location: LOCATION,
+      googleAuth: FAKE_GOOGLE_AUTH,
+    });
+    const chat = model.startChat({
+      generationConfig: {
+        ...TEST_GENERATION_CONFIG,
+        responseMimeType: 'application/json',
+      },
+    });
+
+    expect(chat).toBeInstanceOf(ChatSession);
+  });
   it('set timeout info in ChatSession', () => {
     const model = new GenerativeModel({
       model: 'gemini-pro',
@@ -478,6 +498,22 @@ describe('GenerativeModelPreview startChat', () => {
     });
     const chat = model.startChat({
       history: TEST_USER_CHAT_MESSAGE,
+    });
+
+    expect(chat).toBeInstanceOf(ChatSessionPreview);
+  });
+  it('returns ChatSessionPreview when pass an arg with generationConfig including responseMimeType', async () => {
+    const model = new GenerativeModelPreview({
+      model: 'gemini-pro',
+      project: PROJECT,
+      location: LOCATION,
+      googleAuth: FAKE_GOOGLE_AUTH,
+    });
+    const chat = model.startChat({
+      generationConfig: {
+        ...TEST_GENERATION_CONFIG,
+        responseMimeType: 'application/json',
+      },
     });
 
     expect(chat).toBeInstanceOf(ChatSessionPreview);
@@ -1053,6 +1089,21 @@ describe('GenerativeModel generateContent', () => {
     }
   });
 
+  it('includes responseMimeType', async () => {
+    const reqWithEmptyConfigs: GenerateContentRequest = {
+      contents: TEST_USER_CHAT_MESSAGE_WITH_GCS_FILE,
+      generationConfig: {responseMimeType: 'application/json'},
+      safetySettings: [],
+    };
+    const expectedRequestArgBody =
+      '{"contents":[{"role":"user","parts":[{"text":"How are you doing today?"},{"fileData":{"fileUri":"gs://test_bucket/test_image.jpeg","mimeType":"image/jpeg"}}]}],"generationConfig":{"responseMimeType":"application/json"},"safetySettings":[]}';
+
+    await model.generateContent(reqWithEmptyConfigs);
+    const actualRequestArgsBody = fetchSpy.calls.allArgs()[0][1].body as string;
+
+    expect(actualRequestArgsBody).toEqual(expectedRequestArgBody);
+  });
+
   it('aggregates citation metadata', async () => {
     const req: GenerateContentRequest = {
       contents: TEST_USER_CHAT_MESSAGE,
@@ -1512,6 +1563,21 @@ describe('GenerativeModelPreview generateContent', () => {
     }
   });
 
+  it('includes responseMimeType', async () => {
+    const reqWithEmptyConfigs: GenerateContentRequest = {
+      contents: TEST_USER_CHAT_MESSAGE_WITH_GCS_FILE,
+      generationConfig: {responseMimeType: 'application/json'},
+      safetySettings: [],
+    };
+    const expectedRequestArgsBody =
+      '{"contents":[{"role":"user","parts":[{"text":"How are you doing today?"},{"fileData":{"fileUri":"gs://test_bucket/test_image.jpeg","mimeType":"image/jpeg"}}]}],"generationConfig":{"responseMimeType":"application/json"},"safetySettings":[]}';
+
+    await model.generateContent(reqWithEmptyConfigs);
+    const actualRequestArgsBody = fetchSpy.calls.allArgs()[0][1].body as string;
+
+    expect(actualRequestArgsBody).toEqual(expectedRequestArgsBody);
+  });
+
   it('aggregates citation metadata', async () => {
     const req: GenerateContentRequest = {
       contents: TEST_USER_CHAT_MESSAGE,
@@ -1549,7 +1615,7 @@ describe('GenerativeModelPreview generateContent', () => {
     const expectedBody =
       '{"contents":[{"role":"user","parts":[{"text":"How are you doing today?"}]}],"tools":[{"functionDeclarations":[{"name":"get_current_weather","description":"get weather in a given location","parameters":{"type":"OBJECT","properties":{"location":{"type":"STRING"},"unit":{"type":"STRING","enum":["celsius","fahrenheit"]}},"required":["location"]}}]}]}';
     await model.generateContent(TEST_CHAT_MESSSAGE_TEXT);
-    // @ts-ignore
+
     const actualBody = fetchSpy.calls.allArgs()[0][1].body;
     expect(actualBody).toEqual(expectedBody);
   });
@@ -2219,12 +2285,13 @@ describe('ChatSession', () => {
   });
 
   describe('sendMessage', () => {
+    let fetchSpy: jasmine.Spy;
     const expectedResponse = TEST_MODEL_RESPONSE;
     const fetchResult = Promise.resolve(
       new Response(JSON.stringify(expectedResponse), fetchResponseObj)
     );
     beforeEach(() => {
-      spyOn(global, 'fetch').and.returnValue(fetchResult);
+      fetchSpy = spyOn(global, 'fetch').and.returnValue(fetchResult);
     });
     it('returns a GenerateContentResponse and appends to history', async () => {
       const req = 'How are you doing today?';
@@ -2309,6 +2376,33 @@ describe('ChatSession', () => {
         expectedResourcePath
       );
     });
+
+    it('send a message body with the responseMimeType property to functions', async () => {
+      const expectedResult: GenerateContentResult = {
+        response: TEST_MODEL_RESPONSE,
+      };
+      const model = new GenerativeModel({
+        model: 'gemini-pro',
+        project: PROJECT,
+        location: LOCATION,
+        googleAuth: FAKE_GOOGLE_AUTH,
+      });
+      const chat = model.startChat({
+        generationConfig: {
+          ...TEST_GENERATION_CONFIG,
+          responseMimeType: 'application/json',
+        },
+      });
+      const req = 'How are you doing today?';
+      const expectedBody =
+        '{"contents":[{"role":"user","parts":[{"text":"How are you doing today?"}]}],"generationConfig":{"candidateCount":1,"stopSequences":["hello"],"responseMimeType":"application/json"}}';
+      spyOn(PostFetchFunctions, 'processUnary').and.resolveTo(expectedResult);
+
+      await chat.sendMessage(req);
+      const actualBody = fetchSpy.calls.allArgs()[0][1].body;
+      expect(actualBody).toEqual(expectedBody);
+    });
+
     it('send timeout to functions', async () => {
       const modelWithRequestOptions = new GenerativeModel({
         model: 'gemini-pro',
@@ -2644,6 +2738,7 @@ describe('ChatSessionPreview', () => {
   let chatSessionWithFunctionCall: ChatSessionPreview;
   let model: GenerativeModelPreview;
   let expectedStreamResult: StreamGenerateContentResult;
+  let fetchSpy: jasmine.Spy;
 
   beforeEach(async () => {
     model = new GenerativeModelPreview({
@@ -2667,7 +2762,7 @@ describe('ChatSessionPreview', () => {
     const fetchResult = Promise.resolve(
       new Response(JSON.stringify(expectedStreamResult), fetchResponseObj)
     );
-    spyOn(global, 'fetch').and.returnValue(fetchResult);
+    fetchSpy = spyOn(global, 'fetch').and.returnValue(fetchResult);
   });
 
   describe('sendMessage', () => {
@@ -2755,6 +2850,27 @@ describe('ChatSessionPreview', () => {
       expect(generateContentSpy.calls.allArgs()[0][1]).toEqual(
         expectedResourcePath
       );
+    });
+    it('send a message with responseMimeType to functions', async () => {
+      const model = new GenerativeModelPreview({
+        model: 'gemini-pro',
+        project: PROJECT,
+        location: LOCATION,
+        googleAuth: FAKE_GOOGLE_AUTH,
+      });
+      const chat = model.startChat({
+        generationConfig: {
+          ...TEST_GENERATION_CONFIG,
+          responseMimeType: 'application/json',
+        },
+      });
+      const req = 'How are you doing today?';
+      await chat.sendMessage(req);
+      const expectedBody =
+        '{"contents":[{"role":"user","parts":[{"text":"How are you doing today?"}]}],"generationConfig":{"candidateCount":1,"stopSequences":["hello"],"responseMimeType":"application/json"}}';
+      // @ts-ignore
+      const actualBody = fetchSpy.calls.allArgs()[0][1].body;
+      expect(actualBody).toEqual(expectedBody);
     });
     it('send timeout to functions', async () => {
       const modelWithRequestOptions = new GenerativeModelPreview({